+++ src/components/pages/learn/why/AdditionalFeatures.js
