+++ src/pages/learn/why.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
 import DefaultLayout from 'src/layouts/DefaultNewest'
-import Hero from 'src/components/pages/learn/overview/Hero'
-import ServerlessAppBenefits from 'src/components/pages/learn/overview/AppBenefits'
-import ServerlessFrameworkBenefits from 'src/components/pages/learn/overview/FrameworkBenefits'
-import AdditionalFeatures from 'src/components/pages/learn/overview/AdditionalFeatures'
+import Hero from 'src/components/pages/learn/why/Hero'
+import ServerlessAppBenefits from 'src/components/pages/learn/why/AppBenefits'
+import ServerlessFrameworkBenefits from 'src/components/pages/learn/why/FrameworkBenefits'
+import AdditionalFeatures from 'src/components/pages/learn/why/AdditionalFeatures'
 import { Helmet } from 'src/fragments'
 import prefooter from 'src/components/pages/learn/Prefooter.js'
